backend/requirements.txt
Flask
flask-cors
ultralytics
opencv-python
pygame
python-dotenv
requests
pyttsx3
sounddevice
vosk
pyautogui
groq
smtplib
difflib
numpy
wave
gtts
playsound
SpeechRecognition
